Cleaned up schema file, by separating out mutation functions into own directories
