+++ api/schemas/create_domain/__init__.py
@@ -1,0 +1,1 @@
+from schemas.create_domain.create_domain import CreateDomain
