+++ api/schemas/remove_domain/__init__.py
@@ -1,0 +1,1 @@
+from schemas.remove_domain.remove_domain import RemoveDomain
