+++ api/schemas/remove_organization/__init__.py
@@ -1,0 +1,1 @@
+from schemas.remove_organization.remove_organization import RemoveOrganization
