+++ api/schemas/update_domain/__init__.py
@@ -1,0 +1,1 @@
+from schemas.update_domain.update_domain import UpdateDomain
