+++ api/schemas/update_organization/__init__.py
@@ -1,0 +1,1 @@
+from schemas.update_organization.update_organization import UpdateOrganization
